Safari-611.1.10/LayoutTests/fast/text-autosizing/ios/percent-adjust-percent-line-height-expected.txt

Tests percentage line-height is not adjusted with percentage -webkit-text-size-adjust values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testedLineHeight is expectedLineHeight
PASS successfullyParsed is true

TEST COMPLETE
